New version
This version changes the month of release from str to int (1 to 12).

I added one column "disc" that only prints "1" that refers to the number of discs the album has. This is not standardized on bandcamp but most albums have only 1 disc. This is to fit with my previous excel formatting. Multi-disc releases have to be modified manually.